added a form for changing transaction distribution (#55)
After creating a general transaction, user is redirected to a transaction
details page, where he can precise a distribution. almost-even
distribution is default here.
Also thare is a small change in the resolution algorith: as django-money
is used, when one wants to calculate something, one must use 'amount'
field of a money object